Removed old apis for converting media and now have one. Reworked where the conversion code was located and streamlined events and whatnot.
