+++ API/Extensions/ApplicationServiceExtensions.cs
@@ -50,6 +50,7 @@ public static class ApplicationServiceExtensions
         services.AddScoped<IDeviceService, DeviceService>();
         services.AddScoped<IStatisticService, StatisticService>();
         services.AddScoped<IMediaErrorService, MediaErrorService>();
+        services.AddScoped<IMediaConversionService, MediaConversionService>();
 
         services.AddScoped<IScannerService, ScannerService>();
         services.AddScoped<IMetadataService, MetadataService>();
