+++ openapi.json
@@ -7490,25 +7490,12 @@
         }
       }
     },
-    "/api/Server/convert-bookmarks": {
+    "/api/Server/convert-media": {
       "post": {
         "tags": [
           "Server"
         ],
-        "summary": "Triggers the scheduling of the convert bookmarks job. Only one job will run at a time.",
-        "responses": {
-          "200": {
-            "description": "Success"
-          }
-        }
-      }
-    },
-    "/api/Server/convert-covers": {
-      "post": {
-        "tags": [
-          "Server"
-        ],
-        "summary": "Triggers the scheduling of the convert covers job. Only one job will run at a time.",
+        "summary": "Triggers the scheduling of the convert media job. This will convert all media to the target encoding (except for PNG). Only one job will run at a time.",
         "responses": {
           "200": {
             "description": "Success"
